Add test for variety of date/time stamps
When using this library in Seattle, WA, USA it was found that date
objects with a small time (example: 00:01:00) would be interpreted
by the library as occurring the day before.

This test exposes the problem, but does not offer a solution. It's
assumed the issue is somewhere in toJulian() and toDays(), but
this author does not understand them well enought to venture in.